Use uint64_t for chosen speed.
When parsing the chosen speed from a move, store the intermediate value
as uint64_t rather than unsigned.  We do a range check on it anyway (and
then just allow values up to one million), but we need to make sure the
value gotten from JSON can be stored in the mean time.

+++ src/moveprocessor.cpp
@@ -373,7 +373,7 @@ MaybeSetCharacterSpeed (Character& c, const Json::Value& upd)
       return;
     }
 
-  const unsigned speed = val.asUInt64 ();
+  const uint64_t speed = val.asUInt64 ();
   if (speed == 0 || speed > MAX_CHOSEN_SPEED)
     {
       LOG (WARNING)
